Math App Main Menu Version 1
This is the code for the Cross Platform app's main menu. This was made by keeping all off the pages on the same file, but this is going to end up being a problem, as it will cause the file to get bloated and any debugging or bugs that happen, will be harder fix.

The plan going forward is to make the calculator and the Math game in two separate files, and then use 'main.dart' as the main menu, to act as a bridge between all the pages. That way, it's easier to debug and figure out where bugs are happening.